device-agent/device-agent_build/opt/device-agent/requirements.txt
paho-mqtt
gpiozero
python-dateutil 